+++ KittsInfractionSystem/Features/Enums/InfractionType.cs
@@ -1,14 +1,13 @@
 ﻿namespace KittsInfractionSystem.Features.Enums;
 
-public enum InfractionType
+public enum InfractionType : ushort
 {
-    Other,
-    Warn,
-    Mute,
-    TempMute,
-    Unmute,
-    Kick,
-    Ban,
-    Unban
+    Other = 1,
+    Warn = 2,
+    Mute = 3,
+    TempMute = 4,
+    Unmute = 5,
+    Kick = 6,
+    Ban = 7,
+    Unban = 8
 }
-