+++ KittsInfractionSystem/KittsInfractionSystem.cs
@@ -18,7 +18,7 @@ public class KittsInfractionSystem : Plugin
     public override string Author { get; } = "Kittscloud";
     public override string Description { get; } = "";
     public override LoadPriority Priority { get; } = LoadPriority.High;
-    public override Version Version { get; } = new Version(0, 3, 0);
+    public override Version Version { get; } = new Version(0, 3, 1);
     public override Version RequiredApiVersion { get; } = new Version(LabApiProperties.CompiledVersion);
 
     public static Config Config { get; set; }